hf-upload/requirements.txt
# Hugging Face Spaces Requirements
# FastAPI and Web Framework
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Database - using aiosqlite for serverless (SQLite fallback)
sqlalchemy>=2.0.0
aiosqlite>=0.19.0
asyncpg>=0.29.0

# Vector Database
qdrant-client>=1.7.0

# AI and ML
openai>=1.0.0
anthropic>=0.20.0
google-generativeai>=0.3.0

# Authentication
pyjwt>=2.8.0
passlib[bcrypt]>=1.7.4

# HTTP Client
httpx>=0.25.0

# Utilities
python-dotenv>=1.0.0
python-dateutil>=2.8.0
email-validator>=2.1.0
